python-agents/requirements.txt
crewai>=0.80.0
crewai-tools>=0.15.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
notion-client>=2.2.1
python-dotenv>=1.0.0
openai>=1.0.0
